+++ src/formatters/function.ts
@@ -3,10 +3,9 @@ import util from '../util'
 import { represent } from '../common'
 
 export function formatFunction(this: Function, fmt: Formatter) {
-    if (this.name) {
-        fmt.write('<function ', this.name, '>')
-    } else {
-        fmt.write('<function>')
-    }
+    const value = this.name
+        ? ['<function ', this.name, '>']
+        : '<function>'
+    fmt.write({ style: 'hint', value })
 }
 util.extend(Function, represent, formatFunction)