+++ src/formatters/primitives.ts
@@ -3,47 +3,45 @@ import util from '../util'
 import { represent } from '../common'
 
 export function formatDate(this: Date, fmt: Formatter) {
-    fmt.write('Date(', this.toISOString(), ')')
+    fmt.write({ style: 'date', value: 'Date(' + this.toISOString() + ')' })
 }
 util.extend(Date, represent, formatDate)
 
 export function formatSymbol(this: symbol, fmt: Formatter) {
     const key = Symbol.keyFor(this)
-
-    if (key) {
-        fmt.write('Symbol.for(' + util.escape(key) + ')')
-    } else {
-        fmt.write(this.toString())
-    }
+    const value = key != null
+        ? 'Symbol.for(' + util.escape(key) + ')'
+        : this.toString()
+    fmt.write({ style: 'symbol', value })
 }
 util.extend(Symbol, represent, formatSymbol)
 
 export function formatString(this: string, fmt: Formatter) {
-    fmt.write('"', util.escape(this, '"'), '"')
+    fmt.write({ style: 'string', value: ['"', util.escape(this, '"'), '"'] })
 }
 util.extend(String, represent, formatString)
 
 export function formatRegExp(this: RegExp, fmt: Formatter) {
-    fmt.write('/', this.source, '/', this.flags)
+    fmt.write({ style: 'regexp', value: ['/', this.source, '/', this.flags] })
 }
 util.extend(RegExp, represent, formatRegExp)
 
 export function formatNumberWrapper(this: Boolean | Number, fmt: Formatter) {
-    fmt.write('[' + util.objectName(this)! + ': ' + this.valueOf() + ']')
+    fmt.write({ style: 'number', value: '[' + util.objectName(this)! + ': ' + this.valueOf() + ']' })
 }
 util.extend(Boolean, represent, formatNumberWrapper)
 util.extend(Number, represent, formatNumberWrapper)
 
 export function formatStringWrapper(this: String, fmt: Formatter) {
-    fmt.write('[String: "' + util.escape(this.valueOf(), '"') + '"]')
+    fmt.write({ style: 'string', value: '[String: "' + util.escape(this.valueOf(), '"') + '"]' })
 }
 util.extend(String, represent, formatStringWrapper)
 
 export function formatError(this: Error, fmt: Formatter) {
-    fmt.write('[', this.name)
+    fmt.write({ style: 'hint', value: '[' + this.name })
     if (this.message.length > 0) {
-        fmt.write(': ' + this.message)
+        fmt.write({ style: 'hint', value: ': ' + this.message })
     }
-    fmt.write(']')
+    fmt.write({ style: 'hint', value: ']' })
 }
 util.extend(Error, represent, formatError)